Migrate validate_and_save to Event
